requirements.txt
Flask==2.2.5
flask-cors==3.0.10
PyJWT==2.7.0
